Rename profiler files for consistency
And rearrange includes a bit while here
