+++ 32blit/engine/running_average.hpp
